Verificando Controles en los Roles

+++ app/Livewire/MostrarVacantes.php
@@ -5,6 +5,7 @@ namespace App\Livewire;
 use App\Models\Vacante;
 use Livewire\Component;
 use Illuminate\Support\Facades\Storage;
+use Illuminate\Support\Facades\Gate;
 
 class MostrarVacantes extends Component
 {
@@ -12,11 +13,14 @@ class MostrarVacantes extends Component
    
     public function eliminarVacante(Vacante $vacante)
     {
-        if( $vacante->imagen ) {
-            Storage::delete('public/vacantes/' . $vacante->imagen);            
-        } 
-        
-        $vacante->delete();
+        if (Gate::allows('delete', $vacante)){
+            if( $vacante->imagen ) {
+                Storage::delete('public/vacantes/' . $vacante->imagen);            
+            } 
+            
+            $vacante->delete();
+        }
+           
     }
 
     public function render()
